refactor download client to separte static class
