+++ src/main/kotlin/org/wycliffeassociates/resourcecontainer/media/io/FileDownloadClient.kt
@@ -1,0 +1,56 @@
+package org.wycliffeassociates.resourcecontainer.media.io
+
+import java.io.BufferedInputStream
+import java.io.File
+import java.io.FileOutputStream
+import okhttp3.ResponseBody
+import retrofit2.Call
+import retrofit2.Retrofit
+import retrofit2.http.GET
+import retrofit2.http.Url
+
+interface FileDownloadClient { // implemented by retrofit service
+    @GET
+    fun downloadFile(@Url url: String): Call<ResponseBody>
+}
+
+final class DownloadClient private constructor() {
+    companion object {
+        fun downloadFromUrl(url: String, outputDir: File): File? {
+            val urlFile = File(url)
+            val outputFile = outputDir.resolve(urlFile.name)
+
+            val retrofitService = Retrofit.Builder()
+                .baseUrl(urlFile.parentFile.invariantSeparatorsPath + "/")
+                .build()
+            val downloader: FileDownloadClient = retrofitService.create(
+                FileDownloadClient::class.java
+            )
+
+            val call = downloader.downloadFile(urlFile.name)
+            val response = call.execute()
+
+            if (response.isSuccessful) {
+                val body = response.body()
+                if (body == null) {
+                    println("content not found")
+                } else {
+                    writeTempDownload(body, outputFile)
+                }
+            }
+
+            return if (outputFile.isFile) outputFile else null
+        }
+
+        private fun writeTempDownload(body: ResponseBody, outputFile: File): File {
+            BufferedInputStream(body.byteStream()).use { inputStream ->
+                val bytes = inputStream.readBytes()
+
+                FileOutputStream(outputFile).buffered().use { outputStream ->
+                    outputStream.write(bytes)
+                }
+            }
+            return outputFile
+        }
+    }
+}
